+++ components/Organisms/Live.vue
@@ -18,8 +18,8 @@ section.live
 </template>
 
 <script>
-import Heading from '~/components/Heading.vue'
-import LiveCards from '~/components/LiveCards.vue'
+import Heading from '~/components/Atoms/Heading.vue'
+import LiveCards from '~/components/Molecules/LiveCards.vue'
 import { anchorList, scrollOptions } from '~/assets/js/v-scroll-settings.js'
 
 export default {
