+++ components/Organisms/MusicVideo.vue
@@ -23,7 +23,7 @@ section.music-video
 </template>
 
 <script>
-import Heading from '~/components/Heading.vue'
+import Heading from '~/components/Atoms/Heading.vue'
 
 export default {
   components: {
