+++ components/Organisms/Profile.vue
@@ -14,8 +14,8 @@ section.profile
 </template>
 
 <script>
-import Heading from '~/components/Heading.vue'
-import ProfileDescription from '~/components/ProfileDescription.vue'
+import Heading from '~/components/Atoms/Heading.vue'
+import ProfileDescription from '~/components/Molecules/ProfileDescription.vue'
 
 export default {
   components: {
