+++ components/Organisms/SiteFooter.vue
@@ -12,9 +12,9 @@ footer.site-footer
 </template>
 
 <script>
-import TwitterLink from '~/components/TwitterLink.vue'
-import MailLink from '~/components/MailLink.vue'
-import YoutubeLink from '~/components/YoutubeLink.vue'
+import TwitterLink from '~/components/Atoms/TwitterLink.vue'
+import MailLink from '~/components/Atoms/MailLink.vue'
+import YoutubeLink from '~/components/Atoms/YoutubeLink.vue'
 
 export default {
   components: {
